robot speeds float

+++ main.py
@@ -12,7 +12,7 @@ if __name__ == "__main__":
         print("To much arguments (1- world filename, 2- world size, 3- robot mv_speed, 4- robot rot_speed)")
         exit(1)
 
-    robot, list_balls = init_world(sys.argv[1], int(sys.argv[2]), int(sys.argv[3]), int(sys.argv[4]))
+    robot, list_balls = init_world(sys.argv[1], int(sys.argv[2]), float(sys.argv[3]), float(sys.argv[4]))
 
     graph = init_graph(list_balls + [robot.position], robot)
     print(len(list_balls), len(graph))
